+++ TR2Main-VS/3dsystem/3dinsert.h
@@ -45,7 +45,7 @@ bool InsertObjectEM(short* ptrObj, int vtxCount, D3DCOLOR tint, PHD_UV* em_uv);
 // NOTE: this function is not presented in the original game
 void InsertGourQuad(int x0, int y0, int x1, int y1, int z, D3DCOLOR color0, D3DCOLOR color1, D3DCOLOR color2, D3DCOLOR color3);
 
-float CalculatePolyZ(SORTTYPE sortType, float z0, float z1, float z2, float z3 = -1.0f);
+double CalculatePolyZ(SORTTYPE sortType, double z0, double z1, double z2, double z3 = -1.0);
 BOOL visible_zclip(PHD_VBUF* vtx0, PHD_VBUF* vtx1, PHD_VBUF* vtx2); // 0x00405840
 int ZedClipper(int vtxCount, POINT_INFO* pts, VERTEX_INFO* vtx); // 0x004058B0
 int XYGUVClipper(int vtxCount, VERTEX_INFO* vtx); // 0x004059F0
